% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate Test Output Functions.R
\name{anova_result}
\alias{anova_result}
\title{ANOVA Test}
\usage{
anova_result(group, variable, decimal = 0, p_only = FALSE)
}
\arguments{
\item{group}{the grouping variable}

\item{variable}{the other variable (numeric for ANOVA)}

\item{decimal}{the number of decimal places to be used in output}

\item{p_only}{when FALSE (default), return statistic and p-value; when TRUE
return only the p-value}
}
\description{
\code{anova_result} organizes output from ANOVA test (for 3+ groups) for
publication.
}
