% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression Output Functions.R
\name{calculate_gee_output}
\alias{calculate_gee_output}
\title{GENERALIZED ESTIMATING EQUATION MODELS}
\usage{
calculate_gee_output(model, decimal = 2)
}
\arguments{
\item{model}{the GEE model object}

\item{decimal}{the number of decimal places to be used in output}
}
\description{
\code{calculate_gee_output} organizes GEE model (Gaussian, Binomial, and Poisson)
output into table for publication. Note: Always evaluate output from summary
of model, as well.
}
