% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression Output Functions.R
\name{calculate_glm_output}
\alias{calculate_glm_output}
\title{GENERALIZED LINEAR REGRESSION (GLM)}
\usage{
calculate_glm_output(glm_model)
}
\arguments{
\item{glm_model}{the generalized linear regression model object}

\item{decimal}{the number of decimal places to be used in output}
}
\description{
\code{calculate_glm_output} organizes genearlized linear regression model output
(binomial, quasipoisson, and poisson) into table for publication. Binomial
for binary outcome data; poisson for count outcome data; quasipoisson when
there is overdispersion.Note: Always evaluate output from summary of model,
as well.
}
