% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression Output Functions.R
\name{calculate_glmer_modified_poisson_output}
\alias{calculate_glmer_modified_poisson_output}
\title{GENERALIZED LINEAR MIXED EFFECT MODEL: Modified Poisson}
\usage{
calculate_glmer_modified_poisson_output(model, decimal = 2)
}
\arguments{
\item{model}{the generalized linear mixed effect regression model object}

\item{decimal}{the number of decimal places to be used in output}
}
\description{
\code{calculate_glmer_modified_poisson_output} organizes modified poisson mixed
effect model output into table for publication. Applies standard errors to
be able to use poisson model for binary outcome.
Note: Always evaluate output from summary of model, as well.
}
