% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression Output Functions.R
\name{calculate_glmer_output}
\alias{calculate_glmer_output}
\title{GENERALIZED LINEAR MIXED EFFECT MODEL}
\usage{
calculate_glmer_output(model, decimal = 2)
}
\arguments{
\item{model}{the generalized linear mixed effect regression model object}

\item{decimal}{the number of decimal places to be used in output}
}
\description{
\code{calculate_glmer_output} organizes generalized linear mixed effect
regression model (binomial and poisson) output into table for publication.
Note: Always evaluate output from summary of model, as well.
}
