% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression Output Functions.R
\name{calculate_lm_output}
\alias{calculate_lm_output}
\title{LINEAR REGRESSION}
\usage{
calculate_lm_output(lm_model, decimal = 2)
}
\arguments{
\item{lm_model}{the linear regression model object}

\item{decimal}{the number of decimal places to be used in output}
}
\description{
\code{calculate_lm_output} organizes linear regression model output into table
for publication. Note: Always evaluate output from summary of model, as well.
}
