% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression Output Functions.R
\name{calculate_lmer_output}
\alias{calculate_lmer_output}
\title{LINEAR MIXED EFFECT MODEL}
\usage{
calculate_lmer_output(model, decimal = 0)
}
\arguments{
\item{model}{the linear mixed effect regression model object}

\item{decimal}{the number of decimal places to be used in output}
}
\description{
\code{calculate_lmer_output} organizes linear mixed effect regression model output
into table for publication. Note: Always evaluate output from summary of
model, as well.
}
