% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race and Ethnicity Functions.R
\name{categorize_ethnicity}
\alias{categorize_ethnicity}
\title{Identify Country of Origin as Hispanic or Non-Hispanic}
\usage{
categorize_ethnicity(
  ethnicity_list,
  def = "ombp",
  delimeter = ",",
  file = NULL
)
}
\arguments{
\item{ethnicity_list}{the country of origin data}

\item{def}{the Hispanic Dictionary to use; default is 'ombp' and
other options include 'omb' and 'c'.}

\item{delimeter}{the delimiter the separates the country of origins in each
value of ethnicity_list}

\item{file}{the file pathway of an old ethnicity codebook, only if needed}
}
\description{
\code{categorize_ethnicity} cleans ETHNICITY_LIST and passes the data to another
function called 'hispanic_detailed' to categorize as Hispanic or Non-Hispanic
}
