% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insurance Functions.R
\name{categorize_insurance}
\alias{categorize_insurance}
\title{Categorize Insurance}
\usage{
categorize_insurance(insurance_list)
}
\arguments{
\item{insurance_list}{the vector of patients' insurance. The values must be
formatted as
"Payor - insert_payor_name_here; Plan - insert_benefit_plan_name_here". For
example, "Payor -  AETNA; Plan - AETNA PPO" or
"Payor -  CIGNA; Plan - CIGNA LOCAL PLUS".}
}
\description{
\code{categorize_insurance} categorizes the unique combinations of a patient's
payor and benefit plan as one of the following: Commercial, Medicare,
Medicaid, MassHealth Limited, Auto, Workers' Comp, Other, or Missing.
}
