% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race and Ethnicity Functions.R
\name{categorize_race}
\alias{categorize_race}
\title{Return Factor of a Cleaned Race Variable}
\usage{
categorize_race(race_list, version = "v1", delimeter = ",")
}
\arguments{
\item{race_list}{the initial race list containing these potential values
separated by the specified delimiter: white, black or african american,
asian, native hawaiian or other pacific islander, american indian or alaska
native, hispanic or latino, other, unavailable, declined, patient doesn't
know, or race not listed}

\item{version}{version chooses which version to return. The default "v1"
cleans race such that White + Other for example, returns 'White'.
The other option, "v2", cleans race such that White + Other for example,
returns 'More Than 1 Race'.}

\item{delimeter}{the delimiter the separates the races in each value of
'race_list'}
}
\description{
\code{create_race_list} returns a cleaned race variable of the following
potential values: "White", "Black or African American", "Asian", "American
Indian or Alaska Native", "Native Hawaiian or Other Pacific Islander", "More
Than 1 Race", "Other", "Unknown"
}
