% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc Functions.R
\name{cbind_fill}
\alias{cbind_fill}
\title{Combining datasets (column bind)}
\usage{
cbind_fill(data_left, data_right)
}
\arguments{
\item{data_left}{the first set of columns}

\item{data_right}{the second set of columns}
}
\description{
\code{cbind_fill} Special column bind for combining columns when heach dataset
does not have all the same rows
}
