% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate Test Output Functions.R
\name{chisq_result_from_flag}
\alias{chisq_result_from_flag}
\title{Chi Squared Test (Flag Variable)}
\usage{
chisq_result_from_flag(
  group,
  variable,
  decimal = 0,
  p_only = FALSE,
  mcsim = FALSE
)
}
\arguments{
\item{group}{the grouping variable}

\item{variable}{the other variable (numeric flag variable with 0s and 1s)}

\item{decimal}{the number of decimal places to be used in output}

\item{p_only}{when FALSE (default), return statistic and p-value; when TRUE
return only the p-value}

\item{mcsim}{when FALSE (default), montecarlo simulation is not used; when
TRUE monte carlo simulation is used; this is helpful when small sample}
}
\description{
\code{chisq_result_from_flag} organizes output from chi-squared test for publication,
specifically for variable operationalized as a numeric flag variable with 0s
and 1s.
}
