% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race and Ethnicity Functions.R
\name{clean_binary_ethnicity}
\alias{clean_binary_ethnicity}
\title{Function to Clean Binary Ethnicity Variable}
\usage{
clean_binary_ethnicity(x)
}
\arguments{
\item{x}{the variable "ETHNIC_GROUP_DSC" from EPIC}
}
\description{
\code{clean_binary_ethnicity} takes values of "Not Hispanic", "No Non Hispanic",
"Portugese", "Madagascar", "Hispanic", "Yes Hispanic",
"Prefer not to say/Decline", "Declined", "Unavailable",
"Patient doesn't know", and "NULL" and returns values in (Non-Hispanic,
Hispanic, Declined, and Unavailable)
}
