% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race and Ethnicity Functions.R
\name{clean_list}
\alias{clean_list}
\title{Helper function to format race list}
\usage{
clean_list(x, race_vars, race_label)
}
\arguments{
\item{x}{the race data}

\item{race_vars}{the race variables}

\item{race_label}{the corresponding labels}
}
\description{
\code{clean_list} returns a nicely formatted list of races
}
