% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race and Ethnicity Functions.R
\name{clean_race_variable}
\alias{clean_race_variable}
\title{Helper Function to Categorize Race}
\usage{
clean_race_variable(x, version)
}
\arguments{
\item{x}{the race variable components from 'categorize_race'}

\item{version}{version chooses which version to return. The default "v1"
cleans race such that White + Other for example, returns 'More Than 1 Race'.
The other option, "v2", cleans race such that White + Other for example,
returns 'White'.}
}
\description{
\code{clean_race_variable} Applies the definitions for taking the race components
and assigning one value.
}
