% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race and Ethnicity Functions.R
\name{create_race_list}
\alias{create_race_list}
\title{Return List of Races Separated by Delimiter}
\usage{
create_race_list(race_list, delimeter = ",")
}
\arguments{
\item{race_list}{the initial race list containing these potential values
separated by the specified delimiter: white, black or african american,
asian, native hawaiian or other pacific islander, american indian or alaska
native, hispanic or latino, other, unavailable, declined, patient doesn't
know, or race not listed}

\item{delimeter}{the delimiter the separates the races in each value of
'race_list'}
}
\description{
\code{create_race_list} returns a string that is a list of races (American Indian
or Alaska Native, Asian, Black or African American, Native Hawaiian or Other
Pacific Islander, White, Other, Unknown) separated by the specified delimiter
}
