% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc Functions.R
\name{difference_in_median}
\alias{difference_in_median}
\title{Difference in Median}
\usage{
difference_in_median(group, sample, nits = 500, decimal = 0)
}
\arguments{
\item{group}{the group (only designed for a variable with two levels/
categories)}

\item{sample}{the numeric variable}

\item{nits}{the number of iterations}

\item{decimal}{the number of decimal points to present}
}
\description{
\code{difference_in_median} is a Bootstrap Method to produce 95\% CI of difference
in medians of samples from two groups
}
