% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table 1 Functions.R
\name{frequency_by_patient_group}
\alias{frequency_by_patient_group}
\title{Frequency of Categories of Categorical Variable}
\usage{
frequency_by_patient_group(
  patient_group,
  variable,
  decimal = 0,
  percent_type = "col",
  remove_empty = TRUE,
  use_na = "ifany",
  separated = FALSE
)
}
\arguments{
\item{patient_group}{the grouping variable}

\item{variable}{the categorical variable}

\item{decimal}{the number of decimal places to be used in output}

\item{percent_type}{when "col" (default), it computes column percents; when
"row" it computes row percents}

\item{remove_empty}{when TRUE (default), it removes rows categories with zero}

\item{separated}{TRUE if you want the frequency separate from the percent
and FALSE (default) if you want them pasted into one}
}
\description{
\code{frequency_by_patient_group} returns the frequency and \% of all categories
of a variable (variable) by group (patient_group)
}
