% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table 1 Functions.R
\name{frequency_flag_by_patient_group}
\alias{frequency_flag_by_patient_group}
\title{Frequency of Flag in Binary Variable by Group}
\usage{
frequency_flag_by_patient_group(
  patient_group,
  variable,
  decimal = 0,
  percent_type = "col",
  show_denom = FALSE,
  separated = FALSE
)
}
\arguments{
\item{patient_group}{the grouping variable}

\item{variable}{the flag variable (must be coded as 0s and 1s)}

\item{decimal}{the number of decimal places to be used in output}

\item{percent_type}{when "col" (default), it computes column percents; when
"row" it computes row percents}

\item{show_denom}{when FALSE (default) the output does not include the
denominator; when TRUE the output does include the denominator; note - this
functionality is only built out for when separated = FALSE}

\item{separated}{TRUE if you want the frequency separate from the percent
and FALSE (default) if you want them pasted into one}
}
\description{
\code{frequency_flag_by_patient_group} returns the frequency and \% of the flag (1s) of
a variable (x) coded as 0s and 1s by group
}
