% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table 1 Functions.R
\name{frequency_of_flag_variable}
\alias{frequency_of_flag_variable}
\title{Frequency of Flag in Binary Variable}
\usage{
frequency_of_flag_variable(
  x,
  decimal = 0,
  show_denom = FALSE,
  separated = FALSE
)
}
\arguments{
\item{x}{the flag variable (must be coded as 0s and 1s)}

\item{decimal}{the number of decimal places to be used in output}

\item{show_denom}{when FALSE (default) the output does not include the
denominator; when TRUE the output does include the denominator; note - this
functionality is only built out for when separated = FALSE}

\item{separated}{TRUE if you want the frequency separate from the percent
and FALSE (default) if you want them pasted into one}
}
\description{
\code{frequency_of_flag_variable} returns the frequency and \% of the flag (1s) of
a variable (x) coded as 0s and 1s
}
