% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table 1 Functions.R
\name{frequency_of_value}
\alias{frequency_of_value}
\title{Frequency of Value}
\usage{
frequency_of_value(x, value, decimal = 0, separated = FALSE)
}
\arguments{
\item{x}{the categorical variable}

\item{value}{the category that you want to determine the frequency/percent of}

\item{decimal}{the number of decimal places to be used in output}

\item{separated}{TRUE if you want the frequency separate from the percent
and FALSE (default) if you want them pasted into one}
}
\description{
\code{frequency_of_value} returns the frequency and \% of a specific category
(value) of a variable (x)
}
