% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table 1 Functions.R
\name{frequency_of_variable}
\alias{frequency_of_variable}
\title{Frequency of Categories of Categorical Variable}
\usage{
frequency_of_variable(
  x,
  decimal = 0,
  use_na = "ifany",
  remove_empty = FALSE,
  separated = FALSE
)
}
\arguments{
\item{x}{the categorical variable}

\item{decimal}{the number of decimal places to be used in output}

\item{use_na}{the default ("ifany") shows a row for missing data if there is
any; if set to anything else, it will not.}

\item{remove_empty}{when FALSE (default), this removes the row for a
category in which there are no occurrences; if TRUE, it keeps the row in the
output}

\item{separated}{TRUE if you want the frequency separate from the percent
and FALSE (default) if you want them pasted into one}
}
\description{
\code{frequency_of_variable} returns the frequency and \% of all categories of a
variable (x)
}
