% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc Functions.R
\name{get_filepath_start}
\alias{get_filepath_start}
\title{Get File Path Start}
\usage{
get_filepath_start()
}
\description{
\code{get_filepath_start} is for grabbing the start of file paths to make code
transferrable when moving between PCs.
}
