% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression Output Functions.R
\name{glm_output_for_forest_plot}
\alias{glm_output_for_forest_plot}
\title{GLM (BINOMIAL) output for forest plot}
\usage{
glm_output_for_forest_plot(model, decimal = 2)
}
\arguments{
\item{model}{the logistic regression (GLM) model object}

\item{decimal}{the number of decimal places to be used in output}
}
\description{
\code{glm_output_for_forest_plot} organizes GLM (BINOMIAL/LOGISTIC) output to
help with a forest plot. Note: Always evaluate output from summary of model,
as well.
}
