% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race and Ethnicity Functions.R
\name{hispanic_detailed}
\alias{hispanic_detailed}
\title{Identify Country of Origin as Hispanic or Non-Hispanic}
\usage{
hispanic_detailed(data, definition = "ombp", f)
}
\arguments{
\item{data}{the country of origin data}

\item{definition}{the Hispanic Dictionary to use; default is 'ombp' and
other options include 'omb' and 'c'.}

\item{f}{the file path to another iteration of the codebook}
}
\description{
\code{hispanic_detailed} determines if a country of origin is Hispanic or
Non-Hispanic given a specific definition (one of ombp, omb, or c).
}
