% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate Test Output Functions.R
\name{kruskal_wallis_result}
\alias{kruskal_wallis_result}
\title{KRUSKAL WALLIS}
\usage{
kruskal_wallis_result(group, variable, decimal = 0, p_only = FALSE)
}
\arguments{
\item{group}{the grouping variable}

\item{variable}{the other variable (numeric)}

\item{decimal}{the number of decimal places to be used in output}

\item{p_only}{when FALSE (default), return statistic and p-value; when TRUE
return only the p-value}
}
\description{
\code{kruskal_wallis_result} organizes output from kruskal wallis test (non-parametric
version of ANOVA) for publication.
}
