% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race and Ethnicity Functions.R
\name{return_ethnicity_vars}
\alias{return_ethnicity_vars}
\title{Return cleaned ethnicity helper variables}
\usage{
return_ethnicity_vars(ethnicity_list, delimeter = ",")
}
\arguments{
\item{ethnicity_list}{the country of origin data}

\item{delimeter}{the delimiter the separates the country of origins in each
value of ethnicity_list}
}
\description{
\code{return_ethnicity_vars} returns ethnicity helper variables used in
categorize_ethnicity
}
