% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate Basic ED Data.R
\name{simulate_ed_dataset}
\alias{simulate_ed_dataset}
\title{Simulate ED Dataset}
\usage{
simulate_ed_dataset(n_encounters = 10000, n_patients = 8500, seed = 42)
}
\arguments{
\item{n_encounters}{the number of ED encounters}

\item{n_patients}{the number of patients; needs to be less than the number
of encounters}

\item{seed}{the seed for random number generation}
}
\description{
\code{simulate_ed_dataset} creates a simulated dataset of ED encounters with
variables: encounter_id, patient_id, age, sex, language, esi, and
time_to_abx.
}
