% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc Functions.R
\name{ss_site_cor_adj}
\alias{ss_site_cor_adj}
\title{Sample Size Adjusted for Site-Level Correlation}
\usage{
ss_site_cor_adj(n, nsites, icc)
}
\arguments{
\item{n}{the unadjusted sample size}

\item{nsites}{the number of sites}

\item{icc}{Intraclass Correlation Coefficient (ICC)}
}
\description{
\code{ss_site_cor_adj} is for calculating sample size adjusted for site-level
correlation.
}
