% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table 1 Functions.R
\name{summary_continuous_spread}
\alias{summary_continuous_spread}
\title{Spread of Numeric Variable}
\usage{
summary_continuous_spread(x, decimal = 0, description = deparse(substitute(x)))
}
\arguments{
\item{x}{the continuous variable}

\item{decimal}{the number of decimal places to be used in output}

\item{group}{the grouping variable}
}
\description{
\code{summary_continuous_spread} summarizes a continuous variable with minimum,
Q1, median, Q3, and max
}
