% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table 1 Functions.R
\name{summary_continuous_variable}
\alias{summary_continuous_variable}
\title{Summarize Numeric Variable for Table 1}
\usage{
summary_continuous_variable(
  x,
  measure = "mean",
  decimal = 0,
  description = deparse(substitute(x)),
  range = TRUE,
  separated = FALSE
)
}
\arguments{
\item{x}{the continuous variable}

\item{measure}{can take value "mean" (default) or "median"}

\item{decimal}{the number of decimal places to be used in output}

\item{description}{the name of the variable (x); you can specify your own}

\item{range}{can be TRUE or FALSE; when type = "median", range = TRUE will
output the (Q1, Q3) and range = FALSE will output the interquartile range
value}

\item{separated}{TRUE if you want the mean/median separate from the
SD/IQR/(Q1, Q3) and FALSE (default) if you want them pasted into one}
}
\description{
\code{summary_continuous_variable} summarizes a continuous variable for
a descriptive table.
}
