% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table 1 Functions.R
\name{table_continuous_by_group}
\alias{table_continuous_by_group}
\title{Summarize Numeric Variable By Group for Table 1}
\usage{
table_continuous_by_group(
  group,
  x,
  decimal = 2,
  measure = "mean",
  range = TRUE,
  separated = FALSE
)
}
\arguments{
\item{group}{the grouping variable}

\item{x}{the continuous variable}

\item{decimal}{the number of decimal places to be used in output}

\item{measure}{can take value "mean" (default) or "median"}

\item{range}{can be TRUE or FALSE; when type = "median", range = TRUE will
output the (Q1, Q3) and range = FALSE will output the interquartile range
value}

\item{separated}{TRUE if you want the mean/median separate from the
SD/IQR/(Q1, Q3) and FALSE (default) if you want them pasted into one}
}
\description{
\code{table_continuous_by_group} summarizes a continuous variable by group for
a descriptive table.
}
