% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table 1 Functions.R
\name{table_continuous_by_group_spread}
\alias{table_continuous_by_group_spread}
\title{Spread of Numeric Variable by Group}
\usage{
table_continuous_by_group_spread(group, x, decimal = 2)
}
\arguments{
\item{group}{the grouping variable}

\item{x}{the continuous variable}

\item{decimal}{the number of decimal places to be used in output}
}
\description{
\code{table_continuous_by_group_spread} summarizes a continuous variable by group
with minimum, Q1, median, Q3, and max
}
