% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc Functions.R
\name{update_mmFunctions}
\alias{update_mmFunctions}
\title{Update mmFunctions}
\usage{
update_mmFunctions()
}
\description{
\code{update_mmFunctions} is for installing the latest version of mmFunctions
from github.
}
