% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate Test Output Functions.R
\name{wilcoxon_result}
\alias{wilcoxon_result}
\title{WILCOXON TEST}
\usage{
wilcoxon_result(group, variable, decimal = 0, p_only = FALSE)
}
\arguments{
\item{group}{the grouping variable (two groups)}

\item{variable}{the other variable (numeric)}

\item{decimal}{the number of decimal places to be used in output}

\item{p_only}{when FALSE (default), return statistic and p-value; when TRUE
return only the p-value}
}
\description{
\code{wilcoxon_result} organizes output from wilcoxon test (non-parametric
version of t-test) for publication.
}
